advanced Auth done

+++ controllers/products.js
@@ -3,10 +3,17 @@ const Product = require('../models/product');
 
 exports.deleteProduct = (req, res, next) => {
     //req.session.user.destroyProduct({ where: { id: req.body.id } })
-    Product.findOneAndDelete(req.body.id)
+    Product.findOneAndDelete({ _id: req.body.id, userId: req.user._id })
     .then(result => {
         //console.log(result);
-        res.status(204).redirect('/admin/products');
+        if (result) {
+            //console.log(result);
+            res.status(204).redirect('/admin/products');
+        }
+        else {
+            res.status(401).redirect('/');
+        }
+        
     }).catch(err => {
         console.log(err);
     });
@@ -42,11 +49,19 @@ exports.postEditProduct = (req, res, next) => {
     //req.session.user.updateProduct(param, { where: { id: req.body.id } })
     //const product = new Product(param);
     //product.edit()
-    Product.findOneAndUpdate(req.body.id, param, { new: true})
+    Product.findOneAndUpdate({ _id: req.body.id, userId: req.user._id }, param, { new: true})
     .then(result => {
-            console.log('again update');
-            //console.log(result);
-            res.status(202).redirect('/admin/products');
+            if(result)
+            {
+              console.log('again update');
+              //console.log(result);
+              res.status(202).redirect('/admin/products');  
+            }
+            else
+            {
+                res.status(401).redirect('/');
+            }
+            
         }).catch(err => {
             console.log(err);
         });
@@ -64,16 +79,21 @@ exports.getEditProduct = (req, res, next) => {
     
     const path = '', docTitle = 'Edit Product', script = 'admin/edit-product';
     //req.session.user.getProducts({ where: { id: req.query.id } })
-    Product.findById(req.query.id)
+    Product.findOne({ _id: req.query.id, userId: req.user._id })
     .then(prod => {
         //console.log(req.body.id);
         //console.log(req.query.id);
         console.log(prod);
-        res.render(script, {
-            product: prod,
-            docTitle: docTitle,
-            path: path
-        });
+        if (result) {
+            res.render(script, {
+                product: prod,
+                docTitle: docTitle,
+                path: path
+            });
+        }
+        else {
+            res.status(401).redirect('/');
+        }
     }).catch(err => {
         console.log(err);
     });
@@ -83,7 +103,7 @@ exports.getAdminProduct = (req, res, next) => {
     
     const path = '/admin/products', docTitle = 'Admin Shop', script = 'admin/product-detail';
     //req.session.user.getProducts({ where: { id: req.query.id } })
-    Product.findById(req.query.id).then(prod => {
+    Product.findOne({ _id: req.query.id, userId: req.user._id }).then(prod => {
         //console.log(req.body.id);
         //console.log(req.query.id);
         //console.log(prod);
@@ -101,7 +121,7 @@ exports.getAdminProducts = (req, res, next) => {
     
     const path = '/admin/products', docTitle = 'Admin Shop', script = 'admin/all-products';
     //req.session.user.getProducts()
-    Product.find().then(products => {
+    Product.find({userId: req.user._id}).then(products => {
         //console.log('find all');  
         //console.log(products);
         res.render(script, {
